feat(ticket): restrict unauthorised deleting of tickets

+++ src/features/ticket/actions/delete-ticket.ts
@@ -3,12 +3,33 @@
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { setCookieByKey } from "@/actions/cookies";
-import { fromErrorToActionState } from "@/components/form/utils/to-action-state";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
 
 export const deleteTicket = async (id: string) => {
+  const { user } = await getAuthOrRedirect();
+
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    // if no ticket match or the user isn't the owner of the ticket, return an error state
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState(
+        "ERROR",
+        "Ticket not found or you are not the owner",
+      );
+    }
+
     await prisma.ticket.delete({ where: { id } });
   } catch (error) {
     return fromErrorToActionState(error);
